test: renamed test case
