+++ stories/small_multiples/7_sunbursts.tsx
