+++ stories/small_multiples/small_multiples.stories.tsx
@@ -33,4 +33,4 @@ export { Example as horizontalBars } from './4_horizontal_bars';
 export { Example as gridLines } from './3_grid_lines';
 export { Example as histogramBars } from './5_histogram_bars';
 export { Example as heterogeneous } from './6_heterogeneous_cartesians';
-export { Example as sunbursts } from './7_sunburst_treemap';
+export { Example as sunbursts } from './7_sunbursts';
